fix(PulseMarker): dynamically load marker
(mainly to avoid hydration errors)
