+++ src/components/common/PulseMarker/index.ts
@@ -1,0 +1,3 @@
+import PulseMarker from "./PulseMarker"
+
+export default PulseMarker
